Tests - Combine redundant test classes
Move ExportActionTest into ExportTest
